feat(globe-navigation): implement multi-globe sync toggle button and styles
