+++ src/scripts/components/main/globe-navigation/globe-navigation.module.css
@@ -1,12 +1,50 @@
 .globeNavigation {
   position: absolute;
+  left: 1.375em;
   right: 1.375em;
   bottom: 1.375em;
   display: flex;
-  gap: 0.9375em;
+  align-items: center;
+  justify-content: space-between;
 }
-.globeNavigation > * {
+.globeNavigation > *,
+.navItems > * {
   z-index: 1;
+}
+.navItems {
+  display: flex;
+  gap: 0.9375em;
+  align-items: center;
+  margin-left: auto;
+}
+.toggleMultiGlobeSync {
+  all: unset;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  color: var(--neutral-esa-white);
+  cursor: pointer;
+  background: var(--Supporting-Colours-Deep-Space-1, #3f6c7d);
+  padding: 12px;
+  border-radius: 40px;
+  font-family: NotesEsa, sans-serif;
+  font-size: 16px;
+  font-weight: 700;
+  letter-spacing: 0.32px;
+  z-index: 1;
+}
+@media screen and (max-width: 767px) {
+  .globeNavigation {
+    bottom: 0.9375em;
+  }
+
+  .toggleMultiGlobeSync {
+    aspect-ratio: 1;
+    justify-content: center;
+  }
+  .toggleMultiGlobeSyncLabel {
+    display: none;
+  }
 }
 .projection {
   text-shadow: 0px 0px 2px var(--black);
